chat-service/requirements.txt
fastapi==0.118.3
uvicorn[standard]==0.37.0
python-socketio[asyncio_client]==5.14.1
python-engineio==4.12.3
python-socketio==5.14.1
python-dotenv==1.1.1
pydantic==2.12.0
motor==3.7.1
redis==6.4.0
python-jose==3.5.0
pydantic-settings==2.11.0